requirements.txt
python-telegram-bot==20.3
openai==1.3.0
python-dotenv==1.0.0
pydantic==1.10.13
httpx==0.24.1
certifi==2023.7.22
python-certifi-win32
